requirements.txt
Django==4.2.5
asgiref==3.7.2
django-cors-headers==4.2.0
django-filter==23.3
django-s3-sqlite==0.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.5.0
dj-database-url==1.0.0
psycopg2-binary==2.9.5
python-decouple==3.6
waitress==2.1.2

django-import-export==3.3.1